implement value_counts, with test for DTA tz retention
